restructure modules and files again
