+++ vortex-vector/src/lib.rs
@@ -1,11 +1,10 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: Copyright the Vortex contributors
 
-//! Mutable decompressed (canonical) vectors for Vortex.
-//!
-//! TODO(connor) More docs.
+//! Immutable and mutable decompressed (canonical) vectors for Vortex.
+// TODO(connor): More docs
 
-// TODO(connor)
+// TODO(connor):
 // - Document everything
 // - Figure out correct panic propagation
 // - Figure out exact semantics of `split_off` w.r.t. length of capacity
@@ -19,8 +18,18 @@
 #![deny(missing_docs)]
 
 mod vector;
-pub use vector::ops::{VectorMutOps, VectorOps};
-pub use vector::{Vector, VectorMut};
+pub use vector::Vector;
+
+mod vector_mut;
+pub use vector_mut::VectorMut;
+
+/// Definition and implementation of [`VectorOps`] and [`VectorMutOps`] for [`Vector`] and
+/// [`VectorMut`], respectively.
+///
+/// [`VectorOps`]: ops::VectorOps
+/// [`VectorMutOps`]: ops::VectorMutOps
+mod ops;
+pub use ops::{VectorMutOps, VectorOps};
 
 mod bool;
 mod null;
@@ -29,5 +38,10 @@ mod primitive;
 pub use bool::{BoolVector, BoolVectorMut};
 pub use null::{NullVector, NullVectorMut};
 pub use primitive::{GenericPVector, GenericPVectorMut, PrimitiveVector, PrimitiveVectorMut};
+
+/// Helper macros for working with the different variants of [`Vector`] and [`VectorMut`].
+///
+/// All macros are exported at the crate level with `#[macro_use]`.
+mod macros;
 
 mod private;
